# back-end/openapi.yaml
openapi: 3.0.3
info:
  title: Accounting API
  version: 0.2.0
  description: |
    English: Phase 1 auth endpoints with JWT access/refresh tokens.
    فارسی: فاز ۱؛ اندپوینت‌های احراز هویت با توکن دسترسی و رفرش.
  x-i18n:
    description:
      en: Phase 1 auth endpoints with JWT access/refresh tokens.
      fa: فاز ۱؛ اندپوینت‌های احراز هویت با توکن دسترسی و رفرش.
servers:
  - url: http://localhost:4100/api
paths:
...
  /v1/invoices:
    get:
      summary: List invoices
      x-i18n:
        summary:
          en: List invoices
          fa: فهرست فاکتورها
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of invoices
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Invoice'
                  message:
                    type: string
    post:
      summary: Create invoice (draft)
      x-i18n:
        summary:
          en: Create invoice (draft)
          fa: ایجاد فاکتور (پیش‌نویس)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fiscal_year_id:
                  type: string
                customer_id:
                  type: string
                  nullable: true
                date:
                  type: string
                  format: date
                invoice_no:
                  type: string
                  nullable: true
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/InvoiceItemInput'
              required: [fiscal_year_id, date, items]
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  message:
                    type: string
  /v1/invoices/{id}:
    get:
      summary: Get invoice
      x-i18n:
        summary:
          en: Get invoice
          fa: دریافت فاکتور
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Invoice with items
          content:
            application/json:
              schema:
                type: object
                properties:
                  item:
                    $ref: '#/components/schemas/Invoice'
                  message:
                    type: string
        '404':
          description: Not found
    patch:
      summary: Update invoice (draft only)
      x-i18n:
        summary:
          en: Update invoice (draft only)
          fa: به‌روزرسانی فاکتور (فقط پیش‌نویس)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fiscal_year_id:
                  type: string
                customer_id:
                  type: string
                  nullable: true
                date:
                  type: string
                  format: date
                invoice_no:
                  type: string
                  nullable: true
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/InvoiceItemInput'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  message:
                    type: string
        '404':
          description: Not found
    delete:
      summary: Delete invoice (draft only)
      x-i18n:
        summary:
          en: Delete invoice (draft only)
          fa: حذف فاکتور (فقط پیش‌نویس)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  message:
                    type: string
        '404':
          description: Not found
  /v1/invoices/{id}/post:
    post:
      summary: Post invoice
      x-i18n:
        summary:
          en: Post invoice
          fa: پست کردن فاکتور
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoicePostRequest'
      responses:
        '200':
          description: Posted
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  status:
                    type: string
                    enum: [posted]
                  message:
                    type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Invoice:
      type: object
      properties:
        id:
          type: string
        fiscal_year_id:
          type: string
        customer_id:
          type: string
          nullable: true
        date:
          type: string
          format: date
        invoice_no:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, posted]
        total:
          type: number
        items:
          type: array
          items:
            $ref: '#/components/schemas/InvoiceItem'
      required: [id, fiscal_year_id, date, status]
    InvoiceItem:
      type: object
      properties:
        id:
          type: string
        invoice_id:
          type: string
        product_id:
          type: string
          nullable: true
        quantity:
          type: number
        unit_price:
          type: number
        total:
          type: number
      required: [id, invoice_id, quantity, unit_price, total]
    InvoiceItemInput:
      type: object
      properties:
        product_id:
          type: string
          nullable: true
        quantity:
          type: number
        unit_price:
          type: number
      required: [quantity, unit_price]
    InvoicePostRequest:
      type: object
      properties:
        accounts:
          type: object
          properties:
            receivable_account_id:
              type: string
            sales_account_id:
              type: string
          required: [receivable_account_id, sales_account_id]
        warehouse_id:
          type: string
          nullable: true
      required: [accounts]
...